Extract code for reusability:
- Extract Logo component from navBar Styles file
- Extract Link style into CommonStyles for reusability
- Extract commonly used TRANSITION_TIME constant
